Revert "Add a version of tile reshape that does not cache mappings on device (#33359)"
This reverts commit 2bb1f91ad5f688a2988627e5074e9550c015a1e5.
