+++ ttnn/cpp/ttnn/operations/data_movement/reshape_view/device/device/dataflow/reader_reshape_tiled.cpp
@@ -7,7 +7,7 @@
 #include "dataflow_api.h"
 
 #include "ttnn/operations/data_movement/common/kernels/common.hpp"
-#include "ttnn/operations/data_movement/reshape_view/reshape_kernel_common.hpp"
+#include "ttnn/operations/data_movement/reshape_view/device/hostdevcommon/common.hpp"
 
 using tt::data_movement::common::enhanced_noc_async_read;
 using ttnn::operations::data_movement::reshape::detail::SegmentMapData;
