+++ ttnn/cpp/ttnn/operations/data_movement/reshape_view/device/hostdevcommon/common.hpp
@@ -3,9 +3,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #pragma once
-#include <cstdint>
-#include <array>
-#include <vector>
 
 namespace ttnn::operations::data_movement::reshape::detail {
 
@@ -16,13 +13,8 @@ struct SegmentMapData {
     value_type input_page_offset;
     value_type output_page_offset;
     value_type num_elements;
-    bool operator==(const SegmentMapData& other) const {
-        return input_page_index == other.input_page_index && input_page_offset == other.input_page_offset &&
-               output_page_offset == other.output_page_offset && num_elements == other.num_elements;
-    }
 
     static constexpr uint32_t size = 4;
 };
-inline std::pair<uint16_t, uint16_t> unpack_rt_short(uint32_t packed) { return {packed >> 16, packed & 0xFFFF}; }
 
 }  // namespace ttnn::operations::data_movement::reshape::detail
