--- ttnn/cpp/ttnn/operations/data_movement/reshape_view/device/reshape_tiled_without_device_mapping_pf.hpp
@@ -1,32 +1,0 @@
-// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#pragma once
-
-#include "ttnn/device_operation.hpp"
-#include "ttnn/operations/data_movement/reshape_view/device/reshape_device_operation_types.hpp"
-
-namespace ttnn::operations::data_movement::reshape {
-
-struct ReshapeTiledWithoutDeviceMappingProgramFactory {
-    struct shared_variables_t {
-        tt::tt_metal::KernelHandle reader_kernel_id{};
-        tt::tt_metal::KernelHandle writer_kernel_id{};
-        std::vector<CoreCoord> utilized_cores;
-    };
-    using cached_program_t = ttnn::device_operation::CachedProgram<shared_variables_t>;
-
-    static cached_program_t create(
-        const operation_attributes_t& operation_attributes,
-        const tensor_args_t& tensor_args,
-        tensor_return_value_t& tensor_return_value);
-
-    static void override_runtime_arguments(
-        cached_program_t& cached_program,
-        const operation_attributes_t& operation_attributes,
-        const tensor_args_t& tensor_args,
-        tensor_return_value_t& tensor_return_value);
-};
-
-}  // namespace ttnn::operations::data_movement::reshape
