+++ ttnn/cpp/ttnn/operations/data_movement/reshape_view/reshape.hpp
@@ -36,8 +36,7 @@ struct ReshapeViewOperation {
         const std::optional<MemoryConfig>& memory_config = std::nullopt,
         const std::optional<PadValue>& pad_value = std::nullopt,
         TileReshapeMapMode = TileReshapeMapMode::CACHE,
-        const std::optional<CoreRangeSet>& sub_core_grid = std::nullopt,
-        const std::optional<bool>& on_device_mappings = std::nullopt);
+        const std::optional<CoreRangeSet>& sub_core_grid = std::nullopt);
     static ttnn::Tensor invoke(
         const ttnn::Tensor& input_tensor,
         const ttnn::Shape& logical_shape,
@@ -45,16 +44,14 @@ struct ReshapeViewOperation {
         const std::optional<MemoryConfig>& memory_config = std::nullopt,
         const std::optional<PadValue>& pad_value = std::nullopt,
         TileReshapeMapMode = TileReshapeMapMode::CACHE,
-        const std::optional<CoreRangeSet>& sub_core_grid = std::nullopt,
-        const std::optional<bool>& on_device_mappings = std::nullopt);
+        const std::optional<CoreRangeSet>& sub_core_grid = std::nullopt);
     static ttnn::Tensor invoke(
         const ttnn::Tensor& input_tensor,
         tt::stl::Span<const int32_t> shape_vector,
         const std::optional<MemoryConfig>& memory_config = std::nullopt,
         const std::optional<PadValue>& pad_value = std::nullopt,
         TileReshapeMapMode = TileReshapeMapMode::CACHE,
-        const std::optional<CoreRangeSet>& sub_core_grid = std::nullopt,
-        const std::optional<bool>& on_device_mappings = std::nullopt);
+        const std::optional<CoreRangeSet>& sub_core_grid = std::nullopt);
 };
 
 }  // namespace operations::data_movement
