+++ ttnn/cpp/ttnn/operations/data_movement/reshape_view/reshape_pybind.cpp
@@ -29,16 +29,8 @@ void bind_reshape_view(pybind11::module& module, const data_movement_operation_t
                const std::optional<MemoryConfig>& memory_config,
                const std::optional<PadValue>& pad_value,
                const ttnn::TileReshapeMapMode reshape_tile_mode,
-               const std::optional<CoreRangeSet>& sub_core_grids,
-               const std::optional<bool>& on_device_mappings) -> ttnn::Tensor {
-                return self(
-                    input_tensor,
-                    shape,
-                    memory_config,
-                    pad_value,
-                    reshape_tile_mode,
-                    sub_core_grids,
-                    on_device_mappings);
+               const std::optional<CoreRangeSet>& sub_core_grids) -> ttnn::Tensor {
+                return self(input_tensor, shape, memory_config, pad_value, reshape_tile_mode, sub_core_grids);
             },
             py::arg("input_tensor"),
             py::arg("shape"),
@@ -46,8 +38,7 @@ void bind_reshape_view(pybind11::module& module, const data_movement_operation_t
             py::arg("memory_config") = std::nullopt,
             py::arg("pad_value") = std::nullopt,
             py::arg("reshape_tile_mode") = ttnn::TileReshapeMapMode::CACHE,
-            py::arg("sub_core_grids") = std::nullopt,
-            py::arg("on_device_mappings") = true},
+            py::arg("sub_core_grids") = std::nullopt},
         ttnn::pybind_overload_t{
             [](const data_movement_operation_t& self,
                const ttnn::Tensor& input_tensor,
@@ -56,8 +47,7 @@ void bind_reshape_view(pybind11::module& module, const data_movement_operation_t
                const std::optional<MemoryConfig>& memory_config,
                const std::optional<PadValue>& pad_value,
                const ttnn::TileReshapeMapMode reshape_tile_mode,
-               const std::optional<CoreRangeSet>& sub_core_grids,
-               const std::optional<bool>& on_device_mappings) -> ttnn::Tensor {
+               const std::optional<CoreRangeSet>& sub_core_grids) -> ttnn::Tensor {
                 return self(
                     input_tensor,
                     logical_shape,
@@ -65,8 +55,7 @@ void bind_reshape_view(pybind11::module& module, const data_movement_operation_t
                     memory_config,
                     pad_value,
                     reshape_tile_mode,
-                    sub_core_grids,
-                    on_device_mappings);
+                    sub_core_grids);
             },
             py::arg("input_tensor"),
             py::arg("logical_shape"),
@@ -75,8 +64,7 @@ void bind_reshape_view(pybind11::module& module, const data_movement_operation_t
             py::arg("memory_config") = std::nullopt,
             py::arg("pad_value") = std::nullopt,
             py::arg("reshape_tile_mode") = ttnn::TileReshapeMapMode::CACHE,
-            py::arg("sub_core_grids") = std::nullopt,
-            py::arg("on_device_mappings") = true},
+            py::arg("sub_core_grids") = std::nullopt},
         ttnn::pybind_overload_t{
             [](const data_movement_operation_t& self,
                const ttnn::Tensor& input_tensor,
@@ -84,16 +72,8 @@ void bind_reshape_view(pybind11::module& module, const data_movement_operation_t
                const std::optional<MemoryConfig>& memory_config,
                const std::optional<PadValue>& pad_value,
                const ttnn::TileReshapeMapMode reshape_tile_mode,
-               const std::optional<CoreRangeSet>& sub_core_grids,
-               const std::optional<bool>& on_device_mappings) -> ttnn::Tensor {
-                return self(
-                    input_tensor,
-                    shape,
-                    memory_config,
-                    pad_value,
-                    reshape_tile_mode,
-                    sub_core_grids,
-                    on_device_mappings);
+               const std::optional<CoreRangeSet>& sub_core_grids) -> ttnn::Tensor {
+                return self(input_tensor, shape, memory_config, pad_value, reshape_tile_mode, sub_core_grids);
             },
             py::arg("input_tensor"),
             py::arg("shape"),
@@ -101,8 +81,7 @@ void bind_reshape_view(pybind11::module& module, const data_movement_operation_t
             py::arg("memory_config") = std::nullopt,
             py::arg("pad_value") = std::nullopt,
             py::arg("recreate_mapping_tensor") = ttnn::TileReshapeMapMode::CACHE,
-            py::arg("sub_core_grids") = std::nullopt,
-            py::arg("on_device_mappings") = true});
+            py::arg("sub_core_grids") = std::nullopt});
 }
 }  // namespace detail
 
@@ -130,7 +109,6 @@ void py_bind_reshape_view(pybind11::module& module) {
                 * :attr:`pad_value` (number): Value to pad the output tensor. Default is 0
                 * :attr:`recreate_mapping_tensor` (bool): Advanced option. Set to true to recompute and realloc mapping tensor. This may alleviate DRAM fragmentation but is slow.
                 * :attr:`sub_core_grids` (CoreRangeSet, optional): Specifies sub-core grid ranges for advanced core selection control. Default uses all the cores in the device.
-                * :attr:`on_device_mappings` (bool, optional): Specifies whether to use the tile reshape version with on device mappings. Default is True.
 
 
             Returns:
